Connectivity is now passing files and values correctly.
Updated the configuration settings to pass along the file paths which, by default, are set as relative paths from the base of the data directory.

The JoinedLocationServer now reads folders (or files) of Parquet files, and sticks them into the a location map.
This could be made parallel, but it only takes a minute or so, on my laptop, so probably not worth it.
We may want to consider async loading of the dependent services at some point, because there are definitely things we can do in parallel.