Tile Server now builds an internal map of Centroids
Rather than reading the centroids on each invocation, we now build a map of GEOID->Centroid values on startup (which takes about 10 seconds).

From there, each action is simply a map lookup, which is really fast.

It also means we can get rid of the locking hook, which should reduce contention.

We also no longer need to build/link the abseil string library.