connectivity is now handling the new Safegraph file names
Safegraph changed the naming of their files to now reflect when the data lands, rather than the start date.

This means we need to use our start/end date parameters to make things work correctly.

This really is incredibly brittle, we now need to set the `--start_date` value to match the `date_range_start` value in the file itself.

This just means our processing logic is really brittle and needs to be improved.